mcp-server/requirements.txt
fastapi
uvicorn
sse-starlette
google-cloud-firestore